requirements.txt
Plaintext
flask
python-pptx
gunicorn